fix: moved health route to app-router w. huskyfix
