--- backend/.lintstagedrc.json
@@ -1,5 +1,0 @@
-{
-  "*.ts": [
-    "npm run lint"
-  ]
-}
