--- frontend/.husky/pre-commit
@@ -1,1 +1,0 @@
-yarn run lint
