+++ frontend/package.json
@@ -11,8 +11,7 @@
     "analyze:server": "cross-env BUNDLE_ANALYZE=server next build",
     "analyze:browser": "cross-env BUNDLE_ANALYZE=browser next build",
     "test": "jest --watch",
-    "test:coverage": "jest --watchAll --coverage --silent false",
-    "prepare": "husky"
+    "test:coverage": "jest --watchAll --coverage --silent false"
   },
   "dependencies": {
     "@headlessui/react": "^2.2.0",
@@ -61,7 +60,6 @@
     "eslint": "8.57.1",
     "eslint-config-next": "14.2.23",
     "eslint-config-prettier": "^9.1.0",
-    "husky": "^9.1.7",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-canvas-mock": "^2.5.2",
